requirements.txt
pydantic
python-dotenv
langchain[openai]
langchain-community
langchain-mcp-adapters
langchain-text-splitters
langgraph
langgraph-cli[inmem]
langgraph-supervisor
langgraph-checkpoint-sqlite
langmem
ipynbname
fastmcp
bs4
grandalf
mystmd
